[tests/common] rename reload.py -> config_reload.py
