+++ tests/common/__init__.py
@@ -1,5 +1,5 @@
 from reboot import reboot
-from reload import config_reload
+from config_reload import config_reload
 from port_toggle import port_toggle
 
 __all__ = ['reboot', 'config_reload', 'port_toggle']
